jeugdzorg/requirements.txt
Django==2.0.1
psycopg2==2.7.3.2
Pillow==5.0.0
django-admin-sortable
django-sortedm2m
django-taggit
requests
django-crontab
beautifulsoup4